Refine of the code and improved readability.
