+++ app/javascript/react/App.tsx
@@ -5,40 +5,12 @@ import { CalendarPage } from "./pages/CalendarPage";
 import GlobalStyles from "./assets/styles/global-styles";
 import "./locales/i18n";
 import store from "./store/index";
-import { WeekView } from "./components/DayView/WeekView";
-
-const today = new Date();
-
-const getDayAhead = (ahead: number): Date => {
-  const today = new Date();
-  const upDate = today.setDate(today.getDate() + ahead);
-  return new Date(upDate);
-};
-
-const weeklyData = [
-  { value: 25, date: today },
-  { value: 50, date: getDayAhead(1) },
-  { value: 120, date: getDayAhead(2) },
-  { value: 2, date: getDayAhead(3) },
-  { value: 60, date: getDayAhead(4) },
-  { value: 200, date: getDayAhead(5) },
-  { value: 30, date: getDayAhead(6) },
-];
-
-const colorRanges = {
-  bottom: 0,
-  lower: 40,
-  middle: 60,
-  higher: 80,
-  top: 100,
-};
 
 const App = () => {
   return (
     <Provider store={store}>
       <GlobalStyles />
-      {/* <CalendarPage /> */}
-      <WeekView weeklyData={weeklyData} colorRanges={colorRanges} />
+      <CalendarPage />
     </Provider>
   );
 };
