+++ app/javascript/react/components/WeekView/WeekView.style.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+const HorizontalStack = styled.div`
+  display: flex;
+  padding: 20px;
+  justify-content: space-between;
+  align-items: flex-end;
+  gap: 10px;
+  margin: 10px 10px 0 10px;
+  padding: 10px;
+`;
+
+export {
+  HorizontalStack
+};
