+++ app/javascript/react/pages/CalendarPage.tsx
@@ -6,6 +6,8 @@ import { Navbar } from "../components/Navbar/Navbar";
 import { Graph } from "../components/Graph";
 import { grey100, white } from "../assets/styles/colors";
 import { CalendarStationHeader } from "../components/molecules/CalendarStationHeader/CalendarStationHeader";
+import { WeekView } from "../components/WeekView/WeekView";
+import { colorRanges, weeklyData } from "../components/WeekView/WeeklyMockData";
 
 const PageLayout = styled.div`
   background-color: ${grey100};
@@ -39,6 +41,7 @@ const CalendarPage = () => {
         <StationDataContainer>
           <Graph />
         </StationDataContainer>
+        <WeekView weeklyData={weeklyData} colorRanges={colorRanges} />
       </PageLayout>
     </>
   );
